requirements.txt
discord.py
termcolor
colorama
asyncio